adjust first pass's stack management, add img_tag() and update HTML
prior to this commit the only BBcodes not included in the first pass's stack were BBcodes that had the close_shadow attribute set and did not have a close attribute.  now it's only the second pass that has that restriction.  the problem with it was that it meant that the closing tag of, for example, list was added by close_tags() and not from the tag that the user had entered.  that meant that it didn't decompile correctly.