[ticket/10046] Introduce $phpbb_root_url.
$phpbb_root_path was treated both as a filesystem path (for
things like file inclusion) and as URI path (for generating URLs).
This created a problem when the filesystem path was different
from URI path, such as when register_shutdown_function was
used. Functions running during shutdown sequence execute in
unpredictable working directories, thus to properly refer to
other files it is necessary to have an absolute path in
$phpbb_root_path. Under the old system this would result in
URLs being generated using filesystem path as the prefix, which
is both wrong and an information leak.

Separate filesystem path from URL prefix by adding a new variable
for storing the URL prefix.

PHPBB3-10046