[ticket/6715] Fixed submission path for avatar gallery.
When user submits their selection they submit a hash of category name,
therefore we have to check it against hashes of filesystem directories.

As a side effect of refactoring done it is no longer possible to select
avatars that exist on the file system but are not displayed in the
gallery view.

PHPBB3-6715