[ticket/10933] Create template instance in style class.
Parameters to template constructor are a strict subset of parameters
to style constructor. Therefore instead of creating a template instance
externally and passing it into the style, create the template instance
in the style itself.

This might cause problems with testing. If so we should inject the
class name of the template class and still have the style create
the template instance.

PHPBB3-10933